add response handler, global error handler

+++ controllers/userController.js
@@ -1,63 +1,78 @@
 const { User } = require('../models');
 const hashPassword = require('../utils/hashPassword');
-const { notFound, badRequest } = require('../utils/errorCreator')
+const { notFound, badRequest } = require('../utils/errorCreator');
+const { responseHandler, errorHandler } = require('../utils/responseHandler');
 
 exports.getAllUsers = async (req, res) => {
-  const users = await User.findAll({ attributes: { exclude: ['password'] } });
-  res.json(users);
+  try {
+    const users = await User.findAll({ attributes: { exclude: ['password'] } });
+    if (!users) {
+      throw notFound('users not found');
+    };
+    responseHandler(res, 'All users list:', users);
+  } catch (error) {
+    console.error('get all users error:', error);
+    errorHandler(res, error.code, error.message);
+  };
 };
 
 exports.getUser = async (req, res) => {
-  const userId = req.params.id;
   try {
-    const user = await User.findByPk(userId, { attributes: { exclude: ['password'] } });
+    const userId = req.params.id;
+    if (!userId) {
+      throw badRequest('bad request');
+    };
 
+    const user = await User.findByPk(userId, { attributes: { exclude: ['password'] } });
     if (!user) {
       throw notFound('user not found');
     };
 
-    return res.status(200).json(user);
+    responseHandler(res, 'user data ', user);
   } catch (error) {
     console.error('get user Id error:', error);
-    res.status(error.code).json({ message: `Error: ${error.message}`});
+    errorHandler(res, error.code, error.message);
   };
 };
 
 exports.deleteUser = async (req, res) => {
-  const userId = req.params.id;
   try {
+    const userId = req.params.id;
+    if (!userId) {
+      throw badRequest('bad request');
+    };
     const isUserExist = await User.findByPk(userId);
     if (!isUserExist) {
       throw notFound('user not found');
     };
     await User.destroy({ where: { id: userId } });
-    
-    return res.json(`User with id ${userId} deleted`);
+    responseHandler(res, `User with id ${userId} deleted`);
   } catch (error) {
     console.error('delete error:', error);
-    res.status(error.code).json({ message: `Error: ${error.message}`});
+    errorHandler(res, error.code, error.message);
   };
 };
 
 exports.updateUser = async (req, res) => {
-  if (!req.body) { throw badRequest('missing request body') };
-  const userId = req.params.id;
-
   try {
-    const isUserExist = await User.findByPk(userId);
-    if (!isUserExist) {
+    const userId = req.params.id;
+    if (!userId) {
+      throw badRequest('bad request');
+    };
+    const user = await User.findByPk(userId);
+    if (!user) {
       throw notFound('user not found');
-    }
-
+    };
     if (Object.keys(req.body).includes('password')) {
       const hashedPassword = await hashPassword.hashPassword(req.body.password);
       req.body.password = hashedPassword;
-    }
+    };
+    await User.update(req.body, { where: { id: userId } });
+    const updatedUser = await User.findByPk(userId, { attributes: { exclude: ['password'] } });
 
-    const updatedUser = await User.update(req.body, { where: { id: userId } });
-    return res.json(updatedUser);
+    responseHandler(res, 'user updated', updatedUser);
   } catch (error) {
     console.error('update user error:', error);
-    res.status(error.code).json({ message: `Error: ${error.message}`});
+    errorHandler(res, error.code, error.message);
   };
 };
